feat: create API file/class
Created a class to be the designated API caller to prevent spreading
out network calls.

Tests are skipped, because there isn't much logic.  The class mostly
directly calls Axios' methods.  The only logic is the if/else statement
for errorHandlingWrapper, which is only used to display error messages
to the console.