feat: create App, context, and routes
Implemented App's code, which renders the routes list, the signup and
getUsers methods, and the React context.  The values for the context
are in a React memo, and the methods are in React callbacks, because
they help optimize app performance, as suggested by React.

Also in App, the heading, "Yodlr Design Challenge", is removed from the
view, otherwise it will show up on the sign up and admin pages.

Tests for App will have to be inserted later, after creating the home
page, as the tests would be integration tests.

Implemented the context for user-related methods.

The routes list currently consist of the sign up and admin pages.  The
catch-all route will need to be done after completing the home page,
since the home page would be the destination for it.