chore: move document title assignment
Removed setting the title in index.html and added the title setting in
the sign up and admin pages.  This allows setting different titles for
different pages.

Also added in the sign up page heading that existed in the public
static files of the starter code in the yodlr-take-home git repo.