feat: add styling for errors displayed to users
For the admin page, the error message is now displayed in a Reactstrap
Alert component, instead of being returned in a page of HTML elements
from an if/else statement.  This helps to easily add styling to the
message and reduces duplicated code of the HTML elements.

For the sign up page, added a bottom margin to the submit button for
spacing between it and the error message.